+++ src/app/aboutme/layout.tsx
@@ -1,7 +1,16 @@
+import Link from 'next/link'
+
 export default function DashboardLayout({
   children,
 }: {
   children: React.ReactNode
 }) {
-  return <div>{children}</div>
+  return (
+    <div>
+      <nav>
+        <Link href="/">Home</Link>
+      </nav>
+      {children}
+    </div>
+  )
 }
