--- src/app/signin/layout.tsx
@@ -1,7 +1,0 @@
-export default function DashboardLayout({
-  children,
-}: {
-  children: React.ReactNode
-}) {
-  return <div>{children}</div>
-}
