--- src/app/signin/page.tsx
@@ -1,4 +1,0 @@
-const SignIn = () => {
-  return <div>SignIn</div>
-}
-export default SignIn
